File Upload & Download 15-18 lessons the end

+++ controllers/admin.js
@@ -2,7 +2,7 @@
 const mongoose = require('mongoose')
 const {validationResult} = require('express-validator')
 const Product = require('../models/product')
-
+const fileHelper = require('../util/file')
 
 exports.getAddProduct = (req, res, next) => {
     // if (!req.session.isLoggedIn) {
@@ -201,7 +201,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice
             product.description = updatedDescription
             if(image) {
-
+                fileHelper.deleteFile(product.imageUrl)
                 product.imageUrl = image.path
             }
             return product.save()
@@ -276,9 +276,16 @@ exports.postDeleteProduct = (req, res, next) => {
 
 
     const prodId = req.body.productId
-    // Product.findByIdAndRemove(prodId)
-    Product.deleteOne({_id:prodId, userId: req.user._id })
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'))
+            }
+            fileHelper.deleteFile(product.imageUrl)
+            return Product.deleteOne({_id:prodId, userId: req.user._id })
+        })
         .then(() => {
+            
             console.log('PRODUCT IS DELETED!')
             res.redirect('/admin/products')
         })
@@ -287,6 +294,8 @@ exports.postDeleteProduct = (req, res, next) => {
             error.httpStatusCode = 500
             return next(error)
         })
+    // Product.findByIdAndRemove(prodId)
+        
     // const prodId = req.body.productId
     // Product.deleteById(prodId)
     //     .then(() => {
